admin done

+++ app/Post.php
@@ -5,22 +5,23 @@ namespace App;
 use Illuminate\Database\Eloquent\Model;
 use Cviebrock\EloquentSluggable\Sluggable;
 use Illuminate\Support\Facades\Storage;
+use Carbon\Carbon;
 
 class Post extends Model
 {   
     use Sluggable;
 
-    protected $fillable = ['title', 'content', 'user_id'];
+    protected $fillable = ['title', 'content', 'date'];
 
     const IS_DRAFT = 0;
     const IS_PUBLIC = 1;
 
     public function category(){
-        return $this->hasOne(Category::class);
+        return $this->belongsTo(Category::class);
     }
 
     public function author(){
-        return $this->hasOne(User::class);
+        return $this->belongsTo(User::class, 'user_id');
     }
 
     public function tags(){
@@ -36,7 +37,7 @@ class Post extends Model
         ];
     }
 
-    public static function add($title, $content){
+    public static function add($fields){
         $post = new static;
         $post->fill($fields);
         $post->user_id = 1;
@@ -53,7 +54,9 @@ class Post extends Model
 
     public function remove(){
         
-        Storage::delete('uploads/' . $this->image);
+        if($this->image != null){
+            Storage::delete('uploads/' . $this->image);
+        }
         $this->delete();
     }
 
@@ -62,10 +65,13 @@ class Post extends Model
         if($image == null){
             return ;
         }
-        Storage::delete('uploads/' . $this->image);
+
+        if($this->image != null){
+            Storage::delete('uploads/' . $this->image);
+        }
 
         $filename = str_random(10) . '.' . $image->extension();
-        $image->saveAs('uploads',$filename);
+        $image->storeAs('uploads', $filename);
         $this->image = $filename;
         $this->save();
     }
@@ -74,7 +80,7 @@ class Post extends Model
         if($this->image == null){
             return '/img/no-image.png';
         }
-        return '/uploads' . $this->image;
+        return '/uploads/' . $this->image;
     }
 
 
@@ -93,7 +99,7 @@ class Post extends Model
             return ;
         }
 
-        $this->tags()->sync([$ids]);
+        $this->tags()->sync($ids);
 
 
     }
@@ -134,5 +140,32 @@ class Post extends Model
         return $this->setFeatured();
     }
 
+    public function setDateAttribute($value){
+       $date =  Carbon::createFromFormat('d/m/y', $value)->format('Y-m-d');
+       $this->attributes['date'] = $date; 
+    }
 
+     public function getCategoryTitle(){
+         if($this->category != null){
+             return $this->category->title;
+         }
+
+         return 'Нет категорий';
+     }   
+
+     public function getTagsTitles(){
+        if(!$this->tags->isEmpty()){
+            return implode(', ', $this->tags->pluck('title')->all());
+        }
+
+        return "Нет тегов";
+     }
+
+     public function getDateAttribute($value){
+       $date =  Carbon::createFromFormat('Y-m-d', $value)->format('d/m/y');
+
+       return $date;
+
+
+     }
 }
